+++ src/OpenTelemetry.Exporter.Geneva/TLDExporter/TldLogExporter.cs
@@ -278,7 +278,7 @@ internal sealed class TldLogExporter : TldExporter, IDisposable
                 // before running out of limit instead of STRING_SIZE_LIMIT_CHAR_COUNT.
                 // 2. Trim smarter, by trimming the middle of stack, an
                 // keep top and bottom.
-                var exceptionStack = ToInvariantString(logRecord.Exception);
+                var exceptionStack = logRecord.Exception.ToInvariantString();
                 eb.AddCountedAnsiString("ext_ex_stack", exceptionStack, Encoding.UTF8, 0, Math.Min(exceptionStack.Length, StringLengthLimit));
                 partAFieldsCount++;
             }
@@ -295,6 +295,13 @@ internal sealed class TldLogExporter : TldExporter, IDisposable
         eb.AddCountedString("severityText", logLevels[(int)logLevel]);
         eb.AddUInt8("severityNumber", GetSeverityNumber(logLevel));
         eb.AddCountedAnsiString("name", categoryName, Encoding.UTF8);
+
+        var eventId = logRecord.EventId;
+        if (eventId != default)
+        {
+            eb.AddInt32("eventId", eventId.Id);
+            partBFieldsCount++;
+        }
 
         byte hasEnvProperties = 0;
         bool bodyPopulated = false;
@@ -395,13 +402,6 @@ internal sealed class TldLogExporter : TldExporter, IDisposable
             eb.AddCountedAnsiString("env_properties", serializedEnvPropertiesStringAsBytes, 0, count);
         }
 
-        var eventId = logRecord.EventId;
-        if (eventId != default)
-        {
-            eb.AddInt32("eventId", eventId.Id);
-            partCFieldsCount++;
-        }
-
         eb.SetStructFieldCount(partCFieldsCountPatch, (byte)partCFieldsCount);
     }
 
@@ -482,22 +482,6 @@ internal sealed class TldLogExporter : TldExporter, IDisposable
         }
 
         return result.Slice(0, validNameLength).ToString();
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static string ToInvariantString(Exception exception)
-    {
-        var originalUICulture = Thread.CurrentThread.CurrentUICulture;
-
-        try
-        {
-            Thread.CurrentThread.CurrentUICulture = CultureInfo.InvariantCulture;
-            return exception.ToString();
-        }
-        finally
-        {
-            Thread.CurrentThread.CurrentUICulture = originalUICulture;
-        }
     }
 
     private static readonly Action<LogRecordScope, TldLogExporter> ProcessScopeForIndividualColumns = (scope, state) =>