+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/Implementation/EntityFrameworkInstrumentationEventSource.cs
@@ -16,8 +16,6 @@
 
 using System;
 using System.Diagnostics.Tracing;
-using System.Globalization;
-using System.Threading;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Instrumentation.EntityFrameworkCore.Implementation;
@@ -32,7 +30,7 @@ internal class EntityFrameworkInstrumentationEventSource : EventSource
     {
         if (this.IsEnabled(EventLevel.Error, (EventKeywords)(-1)))
         {
-            this.UnknownErrorProcessingEvent(handlerName, eventName, ToInvariantString(ex));
+            this.UnknownErrorProcessingEvent(handlerName, eventName, ex.ToInvariantString());
         }
     }
 
@@ -75,25 +73,6 @@ internal class EntityFrameworkInstrumentationEventSource : EventSource
         if (this.IsEnabled(EventLevel.Error, EventKeywords.All))
         {
             this.WriteEvent(5, eventName, exception);
-        }
-    }
-
-    /// <summary>
-    /// Returns a culture-independent string representation of the given <paramref name="exception"/> object,
-    /// appropriate for diagnostics tracing.
-    /// </summary>
-    private static string ToInvariantString(Exception exception)
-    {
-        var originalUICulture = Thread.CurrentThread.CurrentUICulture;
-
-        try
-        {
-            Thread.CurrentThread.CurrentUICulture = CultureInfo.InvariantCulture;
-            return exception.ToString();
-        }
-        finally
-        {
-            Thread.CurrentThread.CurrentUICulture = originalUICulture;
         }
     }
 }