+++ src/OpenTelemetry.ResourceDetectors.Azure/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.0.0-alpha.1
+
+Released 2023-Apr-19
+
 * Add AppService resource detector.
 ([#989](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/989))